check_internet_speed/requirements.txt
streamlstreamlit
speedtest-cli
